+++ cmd/all.go
@@ -31,6 +31,7 @@ func init() {
 		cmdMonitor          commands.CmdObjectMonitor
 		cmdPrintConfig      commands.CmdObjectPrintConfig
 		cmdPrintConfigMtime commands.CmdObjectPrintConfigMtime
+		cmdPrintSchedule    commands.CmdObjectPrintSchedule
 		cmdPrintStatus      commands.CmdObjectPrintStatus
 		cmdProvision        commands.CmdObjectProvision
 		cmdSet              commands.CmdObjectSet
@@ -61,6 +62,7 @@ func init() {
 	cmdMonitor.Init(kind, head, &selectorFlag)
 	cmdPrintConfig.Init(kind, subPrint, &selectorFlag)
 	cmdPrintConfigMtime.Init(kind, cmdPrintConfig.Command, &selectorFlag)
+	cmdPrintSchedule.Init(kind, subPrint, &selectorFlag)
 	cmdPrintStatus.Init(kind, subPrint, &selectorFlag)
 	cmdProvision.Init(kind, head, &selectorFlag)
 	cmdSet.Init(kind, head, &selectorFlag)